app/requirement.txt
fastapi==0.110.2
uvicorn== 0.29.0
numpy==1.26.4
tensorflow==2.15.0
Pillow==9.5.0
